refactor(desktop): move dataURLToFile helper into util module
